Moved NNTrainerStruct to own file
