+++ src/NNTrainer.hpp
@@ -5,44 +5,25 @@
 #include "PrintUtilities.hpp"
 #include "NNTrainingData.hpp"
 #include "NNTrainerConfig.hpp"
+#include "NNTrainerStruct.hpp"
 #include <sstream>
 #include <cstddef>
 
-struct NNTrainerStruct: public NNTrainingData, public NNTrainerConfig
-{
-    FeedForwardNeuralNetwork * ffnn; // Storing a pointer to the to-be-trained FFNN
-
-    void copy(NNTrainingData * tdata)
-    {
-        ndata = tdata->ndata;
-        ntraining = tdata->ntraining;
-        xndim = tdata->xndim;
-        yndim = tdata->yndim;
-        x = tdata->x;
-        y = tdata->y;
-        yd1 = tdata->yd1;
-        yd2 = tdata->yd2;
-        w = tdata->w;
-    }
-
-    void copy(NNTrainerConfig * tconfig)
-    {
-        flag_r = tconfig->flag_r;
-        flag_d1 = tconfig->flag_d1;
-        flag_d2 = tconfig->flag_d2;
-        lambda_r = tconfig->lambda_r;
-        lambda_d1 = tconfig->lambda_d1;
-        lambda_d2 = tconfig->lambda_d2;
-    }
-};
 
 class NNTrainer
 {
 protected:
     NNTrainerStruct _tstruct;
-
 public:
-    NNTrainer(NNTrainingData * const tdata, NNTrainerConfig * const tconfig, FeedForwardNeuralNetwork * const ffnn = NULL){_tstruct.copy(tdata); _tstruct.copy(tconfig); _tstruct.ffnn = ffnn;}
+    // construct from individual structures / ffnn
+    NNTrainer(NNTrainingData * const tdata, NNTrainerConfig * const tconfig, FeedForwardNeuralNetwork * const ffnn = NULL)
+    {
+        _tstruct.copyData(tdata); _tstruct.copyConfig(tconfig); _tstruct.ffnn = ffnn;
+    }
+
+    // construct from full trainer struct
+    NNTrainer(NNTrainerStruct * const tstruct){_tstruct = *tstruct;}
+
     virtual ~NNTrainer(){}
 
     // find best fit from a number of nfits fits
